Move admin components into sub directory
