+++ src/main/js/admin/hooks.ts
