+++ src/main/js/index.ts
@@ -24,9 +24,9 @@
 import { binder } from "@scm-manager/ui-extensions";
 import replaceIssueKeys from "./replaceIssueKeys";
 import IssueLinkMarkdownPlugin from "./IssueLinkMarkdownPlugin";
-import IssueTrackerRoute from "./IssueTrackerRoute";
+import IssueTrackerRoute from "./admin/IssueTrackerRoute";
 import {Links} from "@scm-manager/ui-types";
-import IssueTrackerNavLink from "./IssueTrackerNavLink";
+import IssueTrackerNavLink from "./admin/IssueTrackerNavLink";
 
 type PredicateProps = {
   links: Links;
